+++ src/components/DataUploader.vue
@@ -24,7 +24,8 @@
             <v-btn :color="canceled ? 'grey' : 'success'"
                    :disabled="canceled"
                    @click="verify(true)"
-                   block>
+                   block
+                   data-testid="verify_and_send_button">
               <v-icon left dark>{{ icons.mdiCheck }}</v-icon>
               {{ strings['verify_and_send'] }}
             </v-btn>
@@ -122,17 +123,18 @@ import {blobToReadableStream} from 'binconv/dist/src/blobToReadableStream';
 
 import * as openPgpUtils from '@/utils/openpgp-utils';
 import * as pipingUiUtils from "@/piping-ui-utils";
-import {stringsByLang} from "@/strings/strings-by-lang";
+import * as pipingUiRobust from "@/piping-ui-robust";
 import {mdiAlert, mdiCancel, mdiCheck, mdiChevronDown, mdiCloseCircle} from "@mdi/js";
 import type {VerificationStep} from "@/datatypes";
 import VerificationCode from "@/components/VerificationCode.vue";
 import * as pipingUiAuth from "@/piping-ui-auth";
-import {language} from "@/language";
 import {readableBytesString} from "@/utils/readableBytesString";
 import {zipFilesAsBlob} from "@/utils/zipFilesAsBlob";
 import {supportsFetchUploadStreaming} from "@/utils/supportsFetchUploadStreaming";
 import {makePromise} from "@/utils/makePromise";
 import {forceDisableStreamingUpload} from "@/settings/forceDisableStreamingUpload";
+import {useErrorMessage} from "@/useErrorMessage";
+import {strings} from "@/strings/strings";
 
 const props = defineProps<{ composedProps: DataUploaderProps }>();
 
@@ -147,14 +149,7 @@ const progressSetting = ref<{loadedBytes: number, totalBytes?: number}>({
   totalBytes: undefined,
 });
 
-// NOTE: Function makes dynamic language-switch support possible
-//       Delegation is to reassign this value
-const errorMessageDelegate = ref<() => string | Promise<string>>(() => "");
-
-const errorMessage = ref<string>("");
-watch([errorMessageDelegate, language], async () => {
-  errorMessage.value = await errorMessageDelegate.value();
-});
+const {errorMessage, updateErrorMessage} = useErrorMessage();
 
 const xhr: XMLHttpRequest = new XMLHttpRequest();
 const canceled = ref(false);
@@ -184,10 +179,7 @@ const isDoneUpload = computed<boolean>(() => {
 
 const uploadPath = computed<string>(() => urlJoin(props.composedProps.serverUrl, props.composedProps.secretPath));
 
-const hasError = ref<boolean>(false);
-watch(errorMessageDelegate, async () => {
-  hasError.value = await errorMessageDelegate.value() !== "";
-});
+const hasError = computed<boolean>(() => errorMessage.value !== undefined);
 
 const headerIcon = computed<string>(() => {
   if (hasError.value) {
@@ -226,8 +218,6 @@ const isCancelable = computed<boolean>(() =>
   isReadyToUpload && !isDoneUpload.value && !hasError.value && !canceled.value && verificationStep.value.type !== "verification_code_arrived"
 );
 
-// for language support
-const strings = computed(() => stringsByLang(language.value));
 
 const rootElement = ref<Vue>();
 
@@ -254,7 +244,7 @@ onMounted(async () => {
       }
       if (keyExchangeRes.type === 'error') {
         verificationStep.value = {type: 'error'};
-        errorMessageDelegate.value = () => strings.value['key_exchange_error'](keyExchangeRes.errorCode);
+        updateErrorMessage(() => strings.value?.['key_exchange_error'](keyExchangeRes.errorCode));
         return;
       }
       const {key, verificationCode} = keyExchangeRes;
@@ -273,39 +263,69 @@ async function verify(verified: boolean) {
 
   await pipingUiAuth.verify(props.composedProps.serverUrl, props.composedProps.secretPath, key, verified, canceledPromise);
 
+  if (!verified) {
+    return;
+  }
+
   // If verified, send
-  if (verified) {
-    await send(key);
+  const plainBody: Blob = await makePlainBody();
+
+  // Check whether fetch() upload streaming is supported
+  const supportsUploadStreaming = await supportsFetchUploadStreaming(props.composedProps.serverUrl);
+  console.log("streaming upload support: ", supportsUploadStreaming);
+  console.log("force disable streaming upload: ", forceDisableStreamingUpload.value);
+
+  // Convert plain body to ReadableStream
+  const plainStream = blobToReadableStream(plainBody);
+  // Attach progress
+  const plainStreamWithProgress = getReadableStreamWithProgress(plainStream, plainBody.size);
+  // Encrypt
+  const encryptedStream = await openPgpUtils.encrypt(plainStreamWithProgress, key);
+
+  const abortController = new AbortController();
+  canceledPromise.then(() => {
+    abortController.abort();
+  });
+  // Passwordless transfer always uses Piping UI Robust
+  // TODO: notify when canceled because Piping UI Robust on receiver side keeps receiving
+  await pipingUiRobust.sendReadableStream(
+    props.composedProps.serverUrl,
+    props.composedProps.secretPath,
+    encryptedStream,
+    {
+      abortSignal: abortController.signal,
+      fetchUploadStreamingSupported: !forceDisableStreamingUpload.value && supportsUploadStreaming
+    },
+  );
+  return;
+}
+
+async function makePlainBody() {
+  const data: ActualFileObject[] | string = props.composedProps.data;
+  // Text
+  if (typeof data === "string") {
+    return new Blob([data]);
+    // One file
+  } else if (data.length === 1) {
+    return data[0];
+    // Multiple files
+  } else {
+    const files: ActualFileObject[] = data;
+    isCompressing.value = true;
+    // Zip files
+    const zipBlob: Blob = await zipFilesAsBlob(files);
+    isCompressing.value = false;
+    return zipBlob;
   }
 }
 
 async function send(password: string | Uint8Array | undefined) {
-  const data: ActualFileObject[] | string = props.composedProps.data;
-
-  const plainBody: Blob = await (async () => {
-    // Text
-    if (typeof data === "string") {
-      return new Blob([data]);
-      // One file
-    } else if (data.length === 1) {
-      return data[0];
-      // Multiple files
-    } else {
-      const files: ActualFileObject[] = data;
-      isCompressing.value = true;
-      // Zip files
-      const zipBlob: Blob = await zipFilesAsBlob(files);
-      isCompressing.value = false;
-      return zipBlob;
-    }
-  })();
-
+  const plainBody: Blob = await makePlainBody();
   // If password protection is disabled
   if (password === undefined) {
     uploadByXhr(plainBody, plainBody.size);
     return
   }
-
   // Check whether fetch() upload streaming is supported
   const supportsUploadStreaming = await supportsFetchUploadStreaming(props.composedProps.serverUrl);
   console.log("streaming upload support: ", supportsUploadStreaming);
@@ -341,13 +361,17 @@ async function send(password: string | Uint8Array | undefined) {
       duplex: 'half',
       signal: abortController.signal,
     } as RequestInit);
-    // TODO: check res status
+    if (res.status !== 200) {
+      const message = await res.text();
+      updateErrorMessage(() => strings.value?.['fetch_status_error']({status: res.status, message}));
+      return;
+    }
     await res.text();
   } catch (e: any) {
     if (e.name === 'AbortError') {
       return;
     }
-    errorMessageDelegate.value = () => strings.value['data_uploader_xhr_upload_error'];
+    updateErrorMessage(() => strings.value?.['data_uploader_xhr_upload_error']);
   }
 }
 
@@ -373,17 +397,17 @@ function uploadByXhr(body: Blob | Uint8Array, bodyLength: number) {
   };
   xhr.onload = () => {
     if (xhr.status !== 200) {
-      errorMessageDelegate.value = () => strings.value['xhr_status_error']({
+      updateErrorMessage(() => strings.value?.['xhr_status_error']({
         status: xhr.status,
         response: xhr.responseText
-      });
+      }));
     }
   };
   xhr.onerror = (ev) => {
-    errorMessageDelegate.value = () => strings.value['data_uploader_xhr_onerror']({serverUrl: props.composedProps.serverUrl});
+    updateErrorMessage(() => strings.value?.['data_uploader_xhr_onerror']({serverUrl: props.composedProps.serverUrl}));
   };
   xhr.upload.onerror = () => {
-    errorMessageDelegate.value = () => strings.value['data_uploader_xhr_upload_error'];
+    updateErrorMessage(() => strings.value?.['data_uploader_xhr_upload_error']);
   };
   xhr.send(body);
   // Initialize progress bar