+++ src/strings/en.ts
@@ -7,7 +7,6 @@ const sanitizeHtmlAllowingATagAsync = () => import('@/utils/sanitizeHtmlAllowing
 export const en = {
   language: 'Language',
   dark_theme: 'Dark Theme',
-  pwa_update: 'Update',
   version: `Version: ${VERSION}`,
   view_on_github: 'View source code on GitHub',
   open_source_licenses: 'Open source licenses',
@@ -70,7 +69,13 @@ export const en = {
   xhr_status_error: (p: {status: number, response: string}) => {
     return `Error (${p.status}): "${p.response}"`;
   },
-  data_viewer_xhr_onerror: 'Download error',
+  data_viewer_fetch_error: 'Download error',
+  fetch_status_error: (p: {status: number, message: string}) => {
+    return `Error (${p.status}): "${p.message}"`;
+  },
+  data_viewer_body_read_error: (p: { error: unknown }) => {
+    return `Body read error: ${p.error}`;
+  },
   save: 'Save',
   record_server_url: 'Record server URL',
   record_secret_path: 'Record secret path',