+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.12.0] - 2022-12-26
+### Change
+- Update dependencies
+- Robust and streaming transfer for every browser when using passwordless protection
+
+### Added
+- (internal) Add E2E testing
+
+### Fixed
+- Improve HTTP status checking
+
 ## [0.11.0] - 2022-09-20
 ### Change
 - Improve error message in key exchange
@@ -411,7 +422,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - First release
 
-[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.12.0...v0.11.0
 [0.11.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/nwtgck/piping-ui-web/compare/v0.8.0...v0.9.0